Unit tests pass pod lib lint

+++ Example/Database/Tests/Unit/FTrackedQueryManagerTest.m
@@ -195,8 +195,8 @@
     FTrackedQueryManager *manager = [self newManagerWithClock:clock];
 
     for (NSUInteger i = 0; i < 10; i++) {
-        [manager setQueryActive:[FQuerySpec defaultQueryAtPath:PATH(([NSString stringWithFormat:@"%lu", i]))]];
-        [manager setQueryInactive:[FQuerySpec defaultQueryAtPath:PATH(([NSString stringWithFormat:@"%lu", i]))]];
+      [manager setQueryActive:[FQuerySpec defaultQueryAtPath:PATH(([NSString stringWithFormat:@"%lu", (unsigned long)i]))]];
+      [manager setQueryInactive:[FQuerySpec defaultQueryAtPath:PATH(([NSString stringWithFormat:@"%lu", (unsigned long)i]))]];
         [clock tick];
     }
 
@@ -240,8 +240,10 @@
     FTrackedQueryManager *manager = [self newManagerWithClock:clock];
 
     for (NSUInteger i = 0; i < 10; i++) {
-        [manager setQueryActive:[FQuerySpec defaultQueryAtPath:PATH(([NSString stringWithFormat:@"%lu", i]))]];
-        [manager setQueryInactive:[FQuerySpec defaultQueryAtPath:PATH(([NSString stringWithFormat:@"%lu", i]))]];
+        [manager setQueryActive:[FQuerySpec defaultQueryAtPath:
+                                 PATH(([NSString stringWithFormat:@"%lu", (unsigned long)i]))]];
+        [manager setQueryInactive:[FQuerySpec defaultQueryAtPath:
+                                   PATH(([NSString stringWithFormat:@"%lu", (unsigned long)i]))]];
         [clock tick];
     }
